fix: cors 추가 및 수정
